+++ src/utils/pick.js
@@ -1,0 +1,16 @@
+/**
+ * @author recompose (https://github.com/acdlite/recompose)
+ */
+
+const pick = (obj, keys) => {
+  const result = {};
+  for (let i = 0; i < keys.length; i++) {
+    const key = keys[i];
+    if (obj.hasOwnProperty(key)) {
+      result[key] = obj[key];
+    }
+  }
+  return result;
+};
+
+export default pick;